requirements.txt
behave==1.2.6
appium-python-client==5.2.4
selenium==4.30.0
python-dotenv==1.0.0
requests==2.31.0
pytest==8.0.0
